% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare_combine.r
\name{bare_combine}
\alias{bare_combine}
\title{Turn a selection of comma-separated bare strings into a base::combine statement}
\usage{
bare_combine()
}
\description{
Turns \cr\cr
\code{a,b c,d,e f}\cr\cr
or\cr\cr
\code{a, b c, d, e f}\cr\cr
into\cr\cr
\code{c("a", "b c", "d", "e f")}
}
