% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare_rename.r
\name{bare_rename}
\alias{bare_rename}
\title{Turn a selection of comma-separated bare strings into a dplyr::rename statement}
\usage{
bare_rename()
}
\description{
Turns \cr\cr
\code{a,b c,d,e f}\cr\cr
or\cr\cr
\code{a, b c, d, e f}\cr\cr
into\cr\cr
\code{dplyr::rename(a, `b c`, d, `e f`)}
}
