% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare-space-combine.R
\name{bare_space_combine}
\alias{bare_space_combine}
\title{Turn a selection of space-separated bare strings into a base::combine statement}
\usage{
bare_space_combine()
}
\description{
Turns \cr\cr
\code{a b c d e f}\cr\cr
into\cr\cr
\code{c("a", "b", "c", "d", "e", "f")}
}
\details{
If the option `hrbraddins_uspace` is set TRUE, underscores will be
converted to spaces before the final combine statement is created.
}
