% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-package-urls.R
\name{check_package_urls}
\alias{check_package_urls}
\title{Check package URLs}
\usage{
check_package_urls(just_pass_or_fail = FALSE)
}
\arguments{
\item{just_pass_or_fail}{if `TRUE` (default is `FALSE`) just return `TRUE` if
all the URL checks result in HTTP `200` status codes.}
}
\description{
Check package URLs
}
