% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbraddins-package.R
\docType{package}
\name{hrbraddins}
\alias{hrbraddins}
\alias{hrbraddins-package}
\title{Moar Addins for RStudio}
\description{
Provides additional addins for RStudio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/hrbraddins}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/hrbraddins/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
