% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_rows.r
\name{join_rows}
\alias{join_rows}
\title{Join cr/lf-separated selected rows of text into a single space-separated row}
\usage{
join_rows()
}
\description{
Turns
\preformatted{
a
b
c
d
}
into\cr\cr
\code{a b c d}
}
