% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rant.r
\name{rantAddin}
\alias{rantAddin}
\title{Make it easier to annoy followers and reinforce one's entitlement.}
\usage{
rantAddin()
}
\description{
Make it easier to annoy followers and reinforce one's entitlement.
}
