% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh-renviron.R
\name{refresh_renviron}
\alias{refresh_renviron}
\title{Refreshes in-memory environment variables (if it exists)}
\usage{
refresh_renviron()
}
\description{
Equivanent to: `readRenviron("~/.Renviron")`
}
