% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-package-urls.R
\name{check_package_urls}
\alias{check_package_urls}
\title{Check package URLs}
\usage{
check_package_urls()
}
\description{
Check package URLs
}
\note{
This uses non-exported functions from {tools}, one of which
      relies on the `pdftohtml` binary being present on the
      system running this function.
}
